Move dependencies and variable tab to their own files.
